% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_form}
\alias{fetch_form}
\title{Fetch REDCap form}
\usage{
fetch_form(form, .uri = get_redcap_uri(), .token = get_redcap_token())
}
\arguments{
\item{form}{(chr, length 1) The REDCap form to fetch.}

\item{.uri}{(chr, default: get_redcap_uri()) The REDCap URI.}

\item{.token}{(chr, default: get_redcap_token()) The REDCap token.}
}
\value{
(tbl_df) The REDCap form.
}
\description{
Fetch REDCap form
}
\examples{
\donttest{
  fetch_form("followup_postoperatorio_14_30_60_giorno_po") |>
    str(0)
}
}
