% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_redcap_pid}
\alias{get_redcap_pid}
\title{Get the REDCap project ID from the environment.}
\usage{
get_redcap_pid()
}
\value{
(int) The REDCap project ID.
}
\description{
Get the REDCap project ID from the environment.
}
\examples{
get_redcap_pid()
}
