% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_redcap_token}
\alias{get_redcap_token}
\title{Get the REDCap token from the environment.}
\usage{
get_redcap_token()
}
\value{
(chr) The REDCap token, invisibly.
}
\description{
Get the REDCap token from the environment.
}
\examples{
get_redcap_token()
}
