% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_redcap_uri}
\alias{get_redcap_uri}
\title{Get the REDCap URI from the environment.}
\usage{
get_redcap_uri()
}
\value{
(chr) The REDCap URI.
}
\description{
Get the REDCap URI from the environment.
}
\examples{
get_redcap_uri()
}
