% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt_to_tibble.R
\name{gpt_to_tibble}
\alias{gpt_to_tibble}
\title{GPT JSON to tibble}
\usage{
gpt_to_tibble(gpt_json)
}
\arguments{
\item{gpt_json}{(chr) The GPT JSON.}
}
\value{
(tbl_df) The GPT tibble.
}
\description{
GPT JSON to tibble
}
