% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_answer.R
\name{parse_checkbox}
\alias{parse_checkbox}
\title{Parser for checkboxes}
\usage{
parse_checkbox(x)
}
\arguments{
\item{x}{(chr) The character vector to parse.}
}
\value{
(lgl) The parsed logical vector.
}
\description{
Converts GPT answers yes/no to logical.
}
\examples{
parse_checkbox("si")
parse_checkbox("no")
parse_checkbox("NA")
parse_checkbox("N/A")
}
