% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_answer.R
\name{parse_gpt_fctr}
\alias{parse_gpt_fctr}
\title{Parser for GPT factors}
\usage{
parse_gpt_fctr(from_str, to_fct)
}
\arguments{
\item{from_str}{(chr) The character vector to parse.}

\item{to_fct}{(fct) The target factor}
}
\value{
(fct) The parsed factor.
}
\description{
Converts GPT strings reporting categories to factors, using proper
levels derived from the target factor.
}
\note{
The target factor it is supposed that will be overwritten by the
resulting factor of this function.
}
\examples{
parse_gpt_fctr(c("si", "no", "forse"), factor(c("si", "no")))
}
