% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_answer.R
\name{parse_sensazione}
\alias{parse_sensazione}
\title{Parser for  \code{answer} columns}
\usage{
parse_sensazione(x)
}
\arguments{
\item{x}{(chr) The character vector to parse.}
}
\value{
(lgl) The parsed logical vector.
}
\description{
Converts GPT answers yes/no to logical.
}
\examples{
parse_sensazione("si")
parse_sensazione("no")
parse_sensazione("NA")
parse_sensazione("N/A")
}
