% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gpt_on_redcap_instrument.R
\name{query_gpt_on_redcap_instrument}
\alias{query_gpt_on_redcap_instrument}
\title{Query GPT on REDCap instrument}
\usage{
query_gpt_on_redcap_instrument(
  db,
  instrument = c("note_fup", "comments_fup", "details_fup"),
  model = "gpt-4o-mini",
  seed = 1234
)
}
\arguments{
\item{db}{(tbl_df) The REDCap DB for a single form.}

\item{instrument}{(chr) The REDCap instrument of the form to query.}

\item{model}{(chr, default: "gpt-4o-mini") The GPT model to use.}

\item{seed}{(int, default: 1234) The seed for the GPT model.}
}
\value{
(tbl_df) The REDCap DB with GPT responses parsed.
}
\description{
Query GPT on REDCap instrument
}
