% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.gpt-package.R
\docType{package}
\name{redcap.gpt-package}
\alias{redcap.gpt}
\alias{redcap.gpt-package}
\title{redcap.gpt: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/redcap.gpt}
  \item \url{https://ubesp-dctv.github.io/redcap.gpt/}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/redcap.gpt/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Cinzia Anna Maria Papappicco \email{cinziaannamaria.papappicco@ubep.unipd.it} (\href{https://orcid.org/0000-0002-9396-5025}{ORCID})
}

Other contributors:
\itemize{
  \item Giulia Lorenzoni \email{giulia.lorenzoni@ubep.unipd.it} (\href{https://orcid.org/0000-0003-1771-4686}{ORCID}) [contributor]
  \item Dario Gregori \email{drio.gregori@ubep.unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [contributor]
}

}
\keyword{internal}
