% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchers.R
\name{fetch_redcap}
\alias{fetch_redcap}
\title{Fetch REDCap DB}
\usage{
fetch_redcap(
  .uri = get_redcap_uri(),
  .token = get_redcap_token(),
  forms = NULL
)
}
\arguments{
\item{.uri}{(chr, default: get_redcap_uri()) The REDCap URI.}

\item{.token}{(chr, default: get_redcap_token()) The REDCap token.}
}
\value{
(tbl_df) The REDCap DB.
}
\description{
Fetch REDCap DB
}
\examples{
\donttest{
  fetch_redcap() |> 
    str(1)
}
}
